Add vss and vss-agl recipes
Add recipes that use the VSS tools to generate the stock VSS 3.1.1
JSON as well as a version with the extra AGL signals and DBC
configuration for the KUKSA.val CAN feeder.  The latter has the
additional benefit of serving as a working example of using VSS
vspec overlays at build time.

Bug-AGL: SPEC-4761

Change-Id: I6358a8f0a8b64a857dc22cfbd91190ae05546cc0
Signed-off-by: Scott Murray <scott.murray@konsulko.com>