kuksa-val: fix vss-agl dependency for cluster images
Make vss-agl a runtime dependency of the kuksa-val-agl package as the
configuration in that package is what actually requires it, and remove
the previous addition to packagegroup-agl-ivi-services as it will no
longer be necessary.  This fixes the desired KUKSA.val configuration
for e.g. the Flutter cluster image.

Bug-AGL: SPEC-4761

Change-Id: I9ec93acf8e7a542acdb3e50a9cbace50e4928344
Signed-off-by: Scott Murray <scott.murray@konsulko.com>