Add KUKSA.val databroker
Add a recipe for the KUKSA.val databroker, a Rust implementation of
a VSS server with gRPC APIs, and add it to the images that currently
have the C++ kuksa-val server to ease further integration development.
Since the databroker listens for gRPC API requests on a different port
than the C++ server, there is no conflict during this period.  Users
wanting to drive CAN data into the databroker for testing will have to
manually update the kuksa-dbc-feeder configuration during this interim
period.

Changes:
- Bump common KUKSA.val SRCREV to pick up some improvements.
- Add kuksa-databroker recipe to build the databroker.
- Add kuksa-databroker-agl recipe to install AGL-specific configuration
  for the databroker, similar to what has previously been done with
  kuksa-val-agl for the C++ server.
- Add missing branch node declarations in the AGL VSS overlay, as the
  databroker VSS parser is stricter and treats their absence as an
  error.
- Add the kuksa-databroker, kuksa-databroker-agl, and the
  kuksa-databroker-cli packages to the appropriate packagegroups and
  images.  Some follow up work will factor out a packagegroup for the
  KUKSA.val packages to make future changes easier.

Bug-AGL: SPEC-4762

Change-Id: Ie7837b7e42a2283914adc913b75639fcb83029a6
Signed-off-by: Scott Murray <scott.murray@konsulko.com>