Upgrade KUKSA.val components to 0.4.0 releases
Move kuksa.val and kuksa.val.feeders repos to a commit that includes
the recent 0.4.0 release plus a few post-release fixes, and update
our local patches for the new version.  A new patch is required for
the databroker to disable the upstream addition of a vendored build
of protobuf tooling for the host platform, as that does not work
in the OE build enviroment, and is effectively not required since
known versions of protoc and libprotobuf are already present.

The major change coming in is a rework of TLS support, which requires
a minor change to the DBC feeder configuration file that has also
been included.

Bug-AGL: SPEC-4762

Change-Id: I1a43248f15c9946d9d9dc1e8aacf1c3a96ab45a0
Signed-off-by: Scott Murray <scott.murray@konsulko.com>